refactor docs
