+++ src/components/Sidepanel.svelte
@@ -1,0 +1,154 @@
+<script>
+    let { isSidepanelOpen = $bindable(false), index, length, offset, nextIndex, ref } = $props();
+
+    const sections = ['Top', 'Install', 'How to Use', 'Features', 'Quirks'];
+
+    $effect(() => {
+        nextIndex = Math.floor(Math.random() * length);
+    });
+</script>
+
+<div
+    class="sidepanel {isSidepanelOpen ? 'open' : ''}"
+    style={`
+        position: absolute; 
+        z-index: 100; 
+        display: flex; 
+        flex-direction: column;
+    `}
+>
+    <div class="sidepanel-content">
+        <div class="info">
+            <p>
+                Current: {index}
+            </p>
+            <p>Offset: {offset.y.toFixed(2)}</p>
+            <label for="list-size">Number of Items</label>
+            <input id="list-size" bind:value={length} type="number" />
+        </div>
+        {#each sections as section, i}
+            <button
+                onclick={() => {
+                    ref?.goto(i, { offset: { x: 0, y: 32 } });
+                    isSidepanelOpen = false;
+                }}
+                >#{i} {section}
+            </button>
+            {#if i === 0}
+                <br />
+            {/if}
+        {/each}
+        <br />
+        <button
+            onclick={() => {
+                ref?.goto(nextIndex, { offset: { x: 0, y: 32 } });
+                nextIndex = Math.floor(Math.random() * length);
+                isSidepanelOpen = false;
+            }}
+            >Go to Random
+        </button>
+        <button
+            onclick={() => {
+                ref?.goto(length - 3, { offset: { x: 0, y: 32 } });
+                isSidepanelOpen = false;
+            }}
+            >3rd last Item
+        </button>
+        <button
+            onclick={() => {
+                ref?.goto(length - 2, { offset: { x: 0, y: 32 } });
+                isSidepanelOpen = false;
+            }}
+            >2nd last Item
+        </button>
+        <button
+            onclick={() => {
+                ref?.goto(length - 1, { offset: { x: 0, y: 32 } });
+                isSidepanelOpen = false;
+            }}
+            >Last Item
+        </button>
+    </div>
+</div>
+
+<style>
+    .sidepanel {
+        background: white;
+        box-shadow: 2px 0 4px rgba(0, 0, 0, 0.1);
+        transition: transform 0.3s ease;
+        overflow: auto;
+        padding: 0;
+        font-family:
+            system-ui,
+            -apple-system,
+            BlinkMacSystemFont,
+            'Segoe UI',
+            Roboto,
+            Oxygen,
+            Ubuntu,
+            Cantarell,
+            'Open Sans',
+            'Helvetica Neue',
+            sans-serif;
+        height: 100%;
+    }
+
+    @media (max-width: 1280px) {
+        .sidepanel {
+            position: fixed;
+            top: 0;
+            left: 0;
+            transform: translateX(-100%);
+        }
+
+        .sidepanel.open {
+            transform: translateX(0);
+        }
+    }
+
+    .sidepanel-content {
+        display: flex;
+        flex-direction: column;
+        padding: 1rem;
+    }
+
+    .info {
+        color: gray;
+        display: flex;
+        flex-direction: column;
+        font-size: 0.8rem;
+        gap: 4px;
+
+        padding: 8px;
+        padding-top: 4rem;
+
+        p {
+            margin: 0;
+        }
+    }
+
+    label {
+        font-size: 0.75rem;
+        font-weight: 400;
+        color: gray;
+    }
+
+    button {
+        padding: 8px 16px;
+        margin: 8px;
+        border: 2px solid #222;
+        border-radius: 8px;
+        box-shadow: 4px 4px 0px #222;
+        background: white;
+        color: #222;
+        font-size: 1rem;
+
+        cursor: pointer;
+    }
+
+    button:hover {
+        background: #222;
+        color: white;
+        transition: all 0.2s ease-in-out;
+    }
+</style>
