+++ src/routes/+page.svelte
@@ -1,17 +1,34 @@
 <script>
     import MagicScrollbar from '$lib/MagicScrollbar/MagicScrollbar.svelte';
     import MagicScroller from '$lib/MagicScroller.svelte';
-    import Item from './Item.svelte';
+    import Item from '../components/Item.svelte';
+    import Sidepanel from '../components/Sidepanel.svelte';
 
-    let length = $state(5000000000000);
+    const INITIAL_LENGTH = 5000000000000;
+    let length = $state(INITIAL_LENGTH);
     let height = $state(56);
     let index = $state(0);
     let offset = $state({ x: 0, y: 0 });
     let ref = $state();
-    let nextIndex = $state(Math.floor(Math.random() * length));
+    let nextIndex = $state(0);
 
-    const sections = ['Top', 'Install', 'How to Use', 'Features', 'Quirks'];
+    let isSidepanelOpen = $state(false);
+    let innerWidth = $state(0);
+    const MOBILE_BREAKPOINT = 1280;
+
+    const toggleSidepanel = () => {
+        isSidepanelOpen = !isSidepanelOpen;
+    };
 </script>
+
+<svelte:window bind:innerHeight={height} bind:innerWidth />
+
+<!-- Menu Button - only show on mobile -->
+{#if innerWidth <= MOBILE_BREAKPOINT}
+    <button class="menu-button" onclick={toggleSidepanel}>
+        {isSidepanelOpen ? '✕' : '☰'}
+    </button>
+{/if}
 
 {#snippet item(i)}
     <Item index={i} {length} parent={ref} />
@@ -27,55 +44,7 @@
     <div class="thumb"></div>
 {/snippet}
 
-<svelte:window bind:innerHeight={height} />
-
-<div style={`position: absolute; z-index: 100; display: flex; flex-direction: column;`}>
-    <div class="info">
-        <p>
-            Current: {index}
-        </p>
-        <p>Offset: {offset.y.toFixed(2)}</p>
-    </div>
-    <label for="list-size">Number of Items</label>
-    <input id="list-size" bind:value={length} type="number" />
-    {#each sections as section, i}
-        <button
-            onclick={() => {
-                ref?.goto(i, { offset: { x: 0, y: 32 } });
-            }}
-            >#{i} {section}
-        </button>
-        {#if i === 0}
-            <br />
-        {/if}
-    {/each}
-    <br />
-    <button
-        onclick={() => {
-            ref?.goto(nextIndex, { offset: { x: 0, y: 32 } });
-            nextIndex = Math.floor(Math.random() * length);
-        }}
-        >Go to Random
-    </button>
-    <button
-        onclick={() => {
-            ref?.goto(length - 3, { offset: { x: 0, y: 32 } });
-        }}
-        >3rd last Item
-    </button>
-    <button
-        onclick={() => {
-            ref?.goto(length - 2, { offset: { x: 0, y: 32 } });
-        }}
-        >2nd last Item
-    </button>
-    <button
-        onclick={() => {
-            ref?.goto(length - 1, { offset: { x: 0, y: 32 } });
-        }}
-        >Last Item
-    </button>
-</div>
+<Sidepanel bind:isSidepanelOpen {index} {length} {offset} {nextIndex} {ref} />
 <div class="demo-root">
     <MagicScroller
         bind:this={ref}
@@ -106,30 +75,6 @@
             sans-serif;
     }
 
-    label {
-        font-size: 0.75rem;
-        font-weight: 400;
-        color: gray;
-    }
-    button {
-        padding: 8px 16px;
-        margin: 8px;
-        border: 2px solid #222;
-        border-radius: 8px;
-        box-shadow: 4px 4px 0px #222;
-        background: white;
-        color: #222;
-        font-size: 1rem;
-
-        cursor: pointer;
-    }
-
-    button:hover {
-        background: #222;
-        color: white;
-        transition: all 0.2s ease-in-out;
-    }
-
     .demo-root {
         position: absolute;
         width: 100svw;
@@ -141,15 +86,6 @@
         background: whitesmoke;
         top: 0;
         left: 0;
-    }
-
-    .info {
-        font-size: 0.8rem;
-        color: gray;
-
-        p {
-            margin: 0;
-        }
     }
 
     .track {
@@ -169,5 +105,24 @@
         background: #555;
         user-select: none;
         border-radius: 6px;
+    }
+
+    .menu-button {
+        position: fixed;
+        top: 1rem;
+        left: 1rem;
+        z-index: 101;
+        padding: 0.5rem;
+        font-size: 1.5rem;
+        background: white;
+        border: none;
+        border-radius: 100%;
+        height: 2.5rem;
+        width: 2.5rem;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        user-select: none;
+        box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
     }
 </style>
